style(MemberJoinRequest): 주석 수정

+++ weonhyeok/spring-study/src/main/java/spring_study/spring_study/repository/MemberRepository.java
@@ -5,7 +5,7 @@ import spring_study.spring_study.domain.Member;
 
 import java.util.Optional;
 
-// Member클래스의 기본키는 Integer이다.
-public interface MemberRepository extends JpaRepository<Member, Integer> {
+// Member클래스의 기본키는 Long.
+public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByUserName(String userName);
 }
